CredentialsPolicy.Factory -> CredentialsPolicyFactory
