+++ client-runtime/src/test/java/com/microsoft/rest/v2/CredentialsTests.java
@@ -11,7 +11,7 @@ import com.microsoft.rest.v2.credentials.TokenCredentials;
 
 import com.microsoft.rest.v2.http.HttpMethod;
 import com.microsoft.rest.v2.http.HttpPipeline;
-import com.microsoft.rest.v2.policy.CredentialsPolicy;
+import com.microsoft.rest.v2.policy.CredentialsPolicyFactory;
 import com.microsoft.rest.v2.http.HttpRequest;
 import com.microsoft.rest.v2.http.HttpResponse;
 import com.microsoft.rest.v2.http.MockHttpClient;
@@ -46,7 +46,7 @@ public class CredentialsTests {
 
         final HttpPipeline pipeline = HttpPipeline.build(
                 new MockHttpClient(),
-                new CredentialsPolicy.Factory(credentials),
+                new CredentialsPolicyFactory(credentials),
                 auditorFactory);
 
         HttpRequest request = new HttpRequest("basicCredentialsTest", HttpMethod.GET, new URL("http://localhost"));
@@ -73,7 +73,7 @@ public class CredentialsTests {
 
         HttpPipeline pipeline = HttpPipeline.build(
                 new MockHttpClient(),
-                new CredentialsPolicy.Factory(credentials),
+                new CredentialsPolicyFactory(credentials),
                 auditorFactory);
 
         HttpRequest request = new HttpRequest("basicCredentialsTest", HttpMethod.GET, new URL("http://localhost"));
